+++ tests/nws/products/test_spcpts.py
@@ -35,7 +35,7 @@ def test_100606_closed():
     """Test that we can accurately close off an unclosed polygon."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_closed.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 572.878) < 0.01
+    assert abs(outlook.geometry_layers.area - 572.878) < 0.01
 
 
 def test_130607_larger_than_conus():
@@ -43,21 +43,21 @@ def test_130607_larger_than_conus():
     # /products/outlook/archive/2013/day1otlk_20130607_1630.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_conus.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 4.719) < 0.01
+    assert abs(outlook.geometry_layers.area - 4.719) < 0.01
 
 
 def test_issue246():
     """Test a segment that slightly leaks outside the CONUS."""
     prod = parser(get_test_file("SPCPTS/PTSDY2_greatlakes.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "MRGL", 2)
-    assert abs(outlook.geometry.area - 165.72) < 0.01
+    assert abs(outlook.geometry_layers.area - 165.72) < 0.01
 
 
 def test_880324_largerslight():
     """Test that we discard a polygon that is larger than TSTM."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_larger.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 340.537) < 0.01
+    assert abs(outlook.geometry_layers.area - 340.537) < 0.01
 
 
 def test_210501_multipolygon():
@@ -65,7 +65,7 @@ def test_210501_multipolygon():
     # /products/outlook/archive/2021/day1otlk_20210501_1300.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_multipoly.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 238.516) < 0.01
+    assert abs(outlook.geometry_layers.area - 238.516) < 0.01
 
 
 def test_debugdraw():
@@ -85,7 +85,7 @@ def test_issue466_maine2():
     # /products/outlook/archive/2021/day1otlk_20210602_1300.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_maine2.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 469.71) < 0.01
+    assert abs(outlook.geometry_layers.area - 469.71) < 0.01
 
 
 def test_210703_topoerror():
@@ -93,7 +93,7 @@ def test_210703_topoerror():
     # /products/outlook/archive/2021/day1otlk_20210703_2000.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_topo2.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 452.077) < 0.01
+    assert abs(outlook.geometry_layers.area - 452.077) < 0.01
 
 
 def test_210601_hole():
@@ -101,28 +101,28 @@ def test_210601_hole():
     # /products/outlook/archive/2021/day1otlk_20210601_1300.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_hole.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 315.286) < 0.01
+    assert abs(outlook.geometry_layers.area - 315.286) < 0.01
 
 
 def test_210601_last():
     """Test that the last polygon does not dangle in complex logic."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_last.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 287.92) < 0.01
+    assert abs(outlook.geometry_layers.area - 287.92) < 0.01
 
 
 def test_210519_singlepoint():
     """Test that we handle when a single point is in the PTS."""
     prod = parser(get_test_file("SPCPTS/PTSDY2_single.txt"))
     outlook = prod.get_outlook("HAIL", "0.05", 2)
-    assert abs(outlook.geometry.area - 130.139) < 0.01
+    assert abs(outlook.geometry_layers.area - 130.139) < 0.01
 
 
 def test_890526_multi():
     """Test that we can process this PTS."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_multi.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 111.132) < 0.01
+    assert abs(outlook.geometry_layers.area - 111.132) < 0.01
 
 
 def test_210501_day2_west_coast():
@@ -130,7 +130,7 @@ def test_210501_day2_west_coast():
     # https://.../products/outlook/archive/2021/day2otlk_20210501_1730.html
     prod = parser(get_test_file("SPCPTS/PTSDY2_canada.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 2)
-    assert abs(outlook.geometry.area - 351.342) < 0.01
+    assert abs(outlook.geometry_layers.area - 351.342) < 0.01
 
 
 def test_210427_day1_west_coast():
@@ -138,49 +138,53 @@ def test_210427_day1_west_coast():
     # https://.../products/outlook/archive/2021/day1otlk_20210427_1630.html
     prod = parser(get_test_file("SPCPTS/PTSDY1_west.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 359.536) < 0.01
+    assert abs(outlook.geometry_layers.area - 359.536) < 0.01
 
 
 def test_210409_day2_invalid_geom():
     """Test why this outlook bombed for me."""
     prod = parser(get_test_file("SPCPTS/PTSDY2_invalid3.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 2)
-    assert abs(outlook.geometry.area - 212.9588) < 0.01
+    assert abs(outlook.geometry_layers.area - 212.9588) < 0.01
 
 
 def test_three():
     """Test for a three intersection."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_three.txt"))
-    outlook = prod.get_outlook("CATEGORICAL", "MDT", 1)
-    assert abs(outlook.geometry.area - 31.511) < 0.01
+    outlook = prod.get_outlook("CATEGORICAL", "SLGT", 1)
+    slghtall = outlook.geometry_layers.area
+    slghtreal = outlook.geometry.area
+    moderate = prod.get_outlook("CATEGORICAL", "MDT", 1)
+    modreal = moderate.geometry_layers.area
+    assert abs(slghtall - (slghtreal + modreal)) < 0.01
 
 
 def test_sequence():
     """Test for a bad sequence of multipolygons."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_sequence.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "MDT", 1)
-    assert abs(outlook.geometry.area - 28.441) < 0.01
+    assert abs(outlook.geometry_layers.area - 28.441) < 0.01
 
 
 def test_badpoly3():
     """Test that we can get a slight risk from this."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_badpoly3.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 14.532) < 0.01
+    assert abs(outlook.geometry_layers.area - 14.532) < 0.01
 
 
 def test_badpoly2():
     """Test that we can get a slight risk from this."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_badpoly2.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 47.538) < 0.01
+    assert abs(outlook.geometry_layers.area - 47.538) < 0.01
 
 
 def test_badpoly():
     """Test that we don't get a bad polygon out of this."""
     prod = parser(get_test_file("SPCPTS/PTSDY1_badpoly.txt"))
     outlook = prod.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 271.45) < 0.01
+    assert abs(outlook.geometry_layers.area - 271.45) < 0.01
 
 
 def test_nogeom4():
@@ -188,7 +192,7 @@ def test_nogeom4():
     prod = parser(get_test_file("SPCPTS/PTSDY2_nogeom4.txt"))
     # prod.draw_outlooks()
     outlook = prod.get_outlook("CATEGORICAL", "SLGT", 2)
-    assert abs(outlook.geometry.area - 31.252) < 0.01
+    assert abs(outlook.geometry_layers.area - 31.252) < 0.01
 
 
 def test_may3():
@@ -233,9 +237,10 @@ def test_product_id_roundtrip(dbcursor):
     spc = parser(get_test_file("SPCPTS/PTSDY1_maine.txt"))
     spc.sql(dbcursor)
     dbcursor.execute(
-        "SELECT product_id from spc_outlook where day = 1 and "
-        "product_issue = '2017-06-19 05:56+00' and outlook_type = 'C' LIMIT 1"
+        "SELECT product_id from spc_outlooks where day = 1 and "
+        "product_issue = '2017-06-19 05:56+00' and outlook_type = 'C'"
     )
+    assert dbcursor.rowcount == 9
     assert dbcursor.fetchone()[0] == spc.get_product_id()
 
 
@@ -244,7 +249,7 @@ def test_170619_maine():
     # https://.../products/outlook/archive/2017/day1otlk_20170619_1200.html
     spc = parser(get_test_file("SPCPTS/PTSDY1_maine.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 49.058) < 0.01
+    assert abs(outlook.geometry_layers.area - 49.058) < 0.01
 
 
 def test_issue295_geometryfail():
@@ -265,7 +270,7 @@ def test_200602_unpack():
     # https://.../products/outlook/archive/2020/day2otlk_20200602_1730.html
     spc = parser(get_test_file("SPCPTS/PTSDY2_unpack.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 2)
-    assert abs(outlook.geometry.area - 78.7056) < 0.01
+    assert abs(outlook.geometry_layers.area - 78.7056) < 0.01
 
 
 def test_200109_nogeoms():
@@ -273,7 +278,7 @@ def test_200109_nogeoms():
     # https://.../products/outlook/archive/2020/day2otlk_20200109_1730.html
     spc = parser(get_test_file("SPCPTS/PTSDY2_nogeom3.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "ENH", 2)
-    assert abs(outlook.geometry.area - 33.785) < 0.01
+    assert abs(outlook.geometry_layers.area - 33.785) < 0.01
 
 
 def test_190907_invalid():
@@ -281,35 +286,35 @@ def test_190907_invalid():
     # /products/outlook/archive/2019/day1otlk_20190907_1300.html
     spc = parser(get_test_file("SPCPTS/PTSDY1_190907.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 314.761) < 0.01
+    assert abs(outlook.geometry_layers.area - 314.761) < 0.01
 
 
 def test_190905_invalid():
     """Product hit geos issue."""
     spc = parser(get_test_file("SPCPTS/PTSDY1_geos.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 263.61) < 0.01
+    assert abs(outlook.geometry_layers.area - 263.61) < 0.01
 
 
 def test_190903_invalid():
     """Product hit invalid geometry error."""
     spc = parser(get_test_file("SPCPTS/PTSDY2_invalid2.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 2)
-    assert abs(outlook.geometry.area - 343.74) < 0.01
+    assert abs(outlook.geometry_layers.area - 343.74) < 0.01
 
 
 def test_190801_shapely():
     """Product hit shapely assertion error."""
     spc = parser(get_test_file("SPCPTS/PTSDY1_shapelyerror.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 333.678) < 0.01
+    assert abs(outlook.geometry_layers.area - 333.678) < 0.01
 
 
 def test_190625_nogeom2():
     """This hit some error that we need to debug."""
     spc = parser(get_test_file("SPCPTS/PTSDY2_nogeom2.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 2)
-    assert abs(outlook.geometry.area - 11.59) < 0.01
+    assert abs(outlook.geometry_layers.area - 11.59) < 0.01
 
 
 def test_190527_canada():
@@ -317,7 +322,7 @@ def test_190527_canada():
     # https://.../products/outlook/archive/2019/day1otlk_20190528_0100.html
     spc = parser(get_test_file("SPCPTS/PTSDY1_canada.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "MRGL", 1)
-    assert abs(outlook.geometry.area - 118.229) < 0.01
+    assert abs(outlook.geometry_layers.area - 118.229) < 0.01
 
 
 def test_190515_issue117_month():
@@ -331,7 +336,7 @@ def test_190509_marinebounds():
     """SPC Updated marine bounds."""
     spc = parser(get_test_file("SPCPTS/PTSDY1_marine.txt"))
     outlook = spc.get_outlook("HAIL", "0.15", 1)
-    assert abs(outlook.geometry.area - 17.82) < 0.01
+    assert abs(outlook.geometry_layers.area - 17.82) < 0.01
 
 
 def test_190415_elevated():
@@ -339,7 +344,7 @@ def test_190415_elevated():
     spc = parser(get_test_file("SPCPTS/PFWFD1_example.txt"))
     spc.get_jabbers("")
     outlook = spc.get_outlook("FIRE WEATHER CATEGORICAL", "ELEV", 1)
-    assert abs(outlook.geometry.area - 145.64) < 0.01
+    assert abs(outlook.geometry_layers.area - 145.64) < 0.01
     for level in ["IDRT", "SDRT", "ELEV", "CRIT", "EXTM"]:
         outlook = spc.get_outlook("FIRE WEATHER CATEGORICAL", level, 1)
         assert outlook is not None
@@ -349,7 +354,7 @@ def test_180807_idx1_idx2():
     """This Day1 generated an error."""
     spc = parser(get_test_file("SPCPTS/PTSDY1_idx1_idx2.txt"))
     outlook = spc.get_outlook("WIND", "0.05", 1)
-    assert abs(outlook.geometry.area - 37.83) < 0.02
+    assert abs(outlook.geometry_layers.area - 37.83) < 0.02
 
 
 @pytest.mark.parametrize("database", ["postgis"])
@@ -360,7 +365,7 @@ def test_170926_largeenh(dbcursor):
     # Do twice to force a deletion
     spc.sql(dbcursor)
     outlook = spc.get_outlook("CATEGORICAL", "ENH", 1)
-    assert abs(outlook.geometry.area - 17.50) < 0.01
+    assert abs(outlook.geometry_layers.area - 17.50) < 0.01
 
 
 def test_170703_badday3link():
@@ -382,14 +387,14 @@ def test_170612_nullgeom(dbcursor):
     spc = parser(get_test_file("SPCPTS/PTSD48_nullgeom.txt"))
     spc.sql(dbcursor)
     outlook = spc.get_outlook("ANY SEVERE", "0.15", 4)
-    assert abs(outlook.geometry.area - 56.84) < 0.01
+    assert abs(outlook.geometry_layers.area - 56.84) < 0.01
 
 
 def test_170522_nogeom():
     """See why this has an error with no-geom reported"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_nogeom2.txt"))
     outlook = spc.get_outlook("TORNADO", "0.02", 1)
-    assert abs(outlook.geometry.area - 2.90) < 0.01
+    assert abs(outlook.geometry_layers.area - 2.90) < 0.01
 
 
 def test_170518_bad_dbtime():
@@ -408,7 +413,7 @@ def test_170428_large(dbcursor):
     spc.sql(dbcursor)
     spc.get_outlook("TORNADO", "0.10", 1)
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 428.00) < 0.01
+    assert abs(outlook.geometry_layers.area - 428.00) < 0.01
 
 
 @pytest.mark.parametrize("database", ["postgis"])
@@ -438,16 +443,16 @@ def test_051128_invalid(dbcursor):
     spc = parser(get_test_file("SPCPTS/PTSDY1_biggeom2.txt"))
     spc.sql(dbcursor)
     outlook = spc.get_outlook("WIND", "0.05", 1)
-    assert outlook.geometry.is_empty
+    assert outlook.geometry_layers.is_empty
     print("\n".join(spc.warnings))
-    assert len(spc.warnings) == 2
+    assert len(spc.warnings) == 4
 
 
 def test_080731_invalid():
     """Make sure that the SIG wind threshold does not eat the US"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_biggeom.txt"))
     outlook = spc.get_outlook("WIND", "SIGN", 1)
-    assert abs(outlook.geometry.area - 15.823) < 0.01
+    assert abs(outlook.geometry_layers.area - 15.823) < 0.01
 
 
 def test_170411_jabber_error():
@@ -467,9 +472,9 @@ def test_170406_day48_pre2015(dbcursor):
     """Can we parse a pre2015 days 4-8"""
     spc = parser(get_test_file("SPCPTS/PTSD48_pre2015.txt"))
     outlook = spc.get_outlook("ANY SEVERE", "0.15", 4)
-    assert abs(outlook.geometry.area - 73.116) < 0.01
+    assert abs(outlook.geometry_layers.area - 73.116) < 0.01
     outlook = spc.get_outlook("ANY SEVERE", "0.15", 5)
-    assert abs(outlook.geometry.area - 72.533) < 0.01
+    assert abs(outlook.geometry_layers.area - 72.533) < 0.01
     spc.sql(dbcursor)
 
 
@@ -478,7 +483,7 @@ def test_170406_day48(dbcursor):
     """Can we parse a present day days 4-8"""
     spc = parser(get_test_file("SPCPTS/PTSD48.txt"))
     outlook = spc.get_outlook("ANY SEVERE", "0.15", 4)
-    assert abs(outlook.geometry.area - 40.05) < 0.01
+    assert abs(outlook.geometry_layers.area - 40.05) < 0.01
     spc.sql(dbcursor)
     collect = spc.get_outlookcollection(4)
     assert collect.issue == utc(2017, 4, 9, 12)
@@ -490,56 +495,56 @@ def test_170404_nogeom():
     # 26 Sep 2017, we can workaround this now
     spc = parser(get_test_file("SPCPTS/PTSDY1_2002_nogeom.txt"))
     outlook = spc.get_outlook("TORNADO", "0.05", 1)
-    assert abs(outlook.geometry.area - 8.76) < 0.01
+    assert abs(outlook.geometry_layers.area - 8.76) < 0.01
 
 
 def test_170404_2002():
     """Can we parse something from 2002?"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_2002.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 38.614) < 0.01
+    assert abs(outlook.geometry_layers.area - 38.614) < 0.01
 
 
 def test_170329_notimp():
     """Exception was raised parsing this guy"""
     spc = parser(get_test_file("SPCPTS/PTSDY2_notimp.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "MRGL", 2)
-    assert abs(outlook.geometry.area - 110.24) < 0.01
+    assert abs(outlook.geometry_layers.area - 110.24) < 0.01
 
 
 def test_170215_gh23():
     """A marginal for the entire country :/"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_gh23.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "MRGL", 1)
-    assert abs(outlook.geometry.area - 19.63) < 0.01
+    assert abs(outlook.geometry_layers.area - 19.63) < 0.01
 
 
 def test_150622_ptsdy1_topo():
     """PTSDY1_topo.txt"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_topo.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 91.91) < 0.01
+    assert abs(outlook.geometry_layers.area - 91.91) < 0.01
 
 
 def test_150622_ptsdy2():
     """PTSDY2_invalid.txt parsed ok."""
     spc = parser(get_test_file("SPCPTS/PTSDY2_invalid.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 2)
-    assert abs(outlook.geometry.area - 78.14) < 0.01
+    assert abs(outlook.geometry_layers.area - 78.14) < 0.01
 
 
 def test_150622_ptsdy1():
     """PTSDY1_nogeom.txt"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_nogeom.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 95.912) < 0.01
+    assert abs(outlook.geometry_layers.area - 95.912) < 0.01
 
 
 def test_150612_ptsdy1_3():
     """We got an error with this, so we shall test"""
     spc = parser(get_test_file("SPCPTS/PTSDY1_3.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 1)
-    assert abs(outlook.geometry.area - 53.94) < 0.01
+    assert abs(outlook.geometry_layers.area - 53.94) < 0.01
 
 
 def test_141022_newcats():
@@ -549,23 +554,23 @@ def test_141022_newcats():
         utcnow=utc(2014, 10, 13, 16, 21),
     )
     outlook = spc.get_outlook("CATEGORICAL", "ENH", 1)
-    assert abs(outlook.geometry.area - 13.02) < 0.01
+    assert abs(outlook.geometry_layers.area - 13.02) < 0.01
     outlook = spc.get_outlook("CATEGORICAL", "MRGL", 1)
-    assert abs(outlook.geometry.area - 47.01) < 0.01
+    assert abs(outlook.geometry_layers.area - 47.01) < 0.01
 
 
 def test_140709_nogeoms():
     """Can we parse holes."""
     spc = parser(get_test_file("SPCPTS/PTSDY3_nogeoms.txt"))
     outlook = spc.get_outlook("ANY SEVERE", "0.05", 3)
-    assert abs(outlook.geometry.area - 99.68) < 0.01
+    assert abs(outlook.geometry_layers.area - 99.68) < 0.01
 
 
 def test_140710_nogeom():
     """Can we parse holes."""
     spc = parser(get_test_file("SPCPTS/PTSDY2_nogeom.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "SLGT", 2)
-    assert abs(outlook.geometry.area - 43.02) < 0.01
+    assert abs(outlook.geometry_layers.area - 43.02) < 0.01
 
 
 def test_23jul_failure():
@@ -582,14 +587,14 @@ def test_140707_general():
     # /products/outlook/archive/2014/day1otlk_20140707_1630.html
     spc = parser(get_test_file("SPCPTS/PTSDY1_complex.txt"))
     outlook = spc.get_outlook("CATEGORICAL", "TSTM", 1)
-    assert abs(outlook.geometry.area - 606.333) < 0.01
+    assert abs(outlook.geometry_layers.area - 606.333) < 0.01
 
 
 def test_complex():
     """Test our processing"""
     spc = parser(get_test_file("SPCPTS/PTSDY3.txt"))
     outlook = spc.get_outlook("ANY SEVERE", "0.05", 3)
-    assert abs(outlook.geometry.area - 10.12) < 0.01
+    assert abs(outlook.geometry_layers.area - 10.12) < 0.01
 
 
 def test_bug_140601_pfwf38():
@@ -653,7 +658,7 @@ def test_complex_2():
     """Test our processing"""
     spc = parser(get_test_file("SPCPTS/PTSDY1.txt"))
     outlook = spc.get_outlook("HAIL", "0.05", 1)
-    assert abs(outlook.geometry.area - 47.65) < 0.01
+    assert abs(outlook.geometry_layers.area - 47.65) < 0.01
 
 
 @pytest.mark.parametrize("database", ["postgis"])